requirements.txt
appdirs==1.4.3
Django==1.9
django-easy-pdf==0.1.0
html5lib==1.0b10
httplib2==0.10.3
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
PyPDF2==1.26.0
reportlab==3.4.0
six==1.10.0
webencodings==0.5.1
xhtml2pdf==0.2b1
